requirements.txt
alembic==0.9.1
appdirs==1.4.3
blinker==1.4
click==6.7
decorator==4.0.11
Flask==0.12
Flask-Cache==0.13.1
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Redis==0.3.0
Flask-Script==2.0.5
Flask-SocketIO==2.8.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
pbr==2.0.0
psycopg2==2.7.1
pyparsing==2.2.0
python-editor==1.0.3
python-engineio==1.3.0
python-socketio==1.7.2
redis==2.10.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.1
WTForms==2.1
XlsxWriter==0.9.6
xpinyin==0.5.5
